Fix nightly CI test failures from to_torch unsigned dtype change
PR #36660 changed ttnn.to_torch() to return proper unsigned PyTorch
types (torch.uint16, torch.uint32) instead of silently converting to
signed variants. That PR updated ~20 test files but missed several
nightly tests since they weren't exercised by post-commit CI.

Fixes:
- test_pad.py: cast output to match input dtype for uint16 comparison
- test_embedding.py: cast indices to torch.long for embedding lookup
- test_isin.py: cast uint32 result to int32 before count_nonzero

Made-with: Cursor

+++ tests/ttnn/unit_tests/operations/data_movement/test_embedding.py
@@ -13,7 +13,7 @@ def test_base_case(device):
     torch.manual_seed(1234)
     indices = ttnn.to_device(ttnn.from_torch(torch.tensor([[1, 2, 4, 5], [4, 3, 2, 9]]), dtype=ttnn.uint32), device)
     embedding_matrix = ttnn.to_device(ttnn.from_torch(torch.rand(10, 2), dtype=ttnn.bfloat16), device)
-    indices_torch = ttnn.to_torch(ttnn.from_device(indices))
+    indices_torch = ttnn.to_torch(ttnn.from_device(indices)).to(torch.long)
     embedding_matrix_torch = ttnn.to_torch(ttnn.from_device(embedding_matrix))
     expected_embeddings = torch.nn.functional.embedding(indices_torch, embedding_matrix_torch)
     embeddings = ttnn.embedding(indices, embedding_matrix)
